barrier: Correct gfid in statedump of barriered fops
In brick statedump file the barriered fop's gfid was showing 0 when
statedump was taken. This is because of statedump code was not
referring to correct gfid.

With this change statedump code will use correct gfid and gfid will
not be 0 in statedump file when barrier is enable and user takes
statedump of volume.

Change-Id: Ia296cba7e132402df53c602daa160c1c2cd21245
BUG: 1099369